Fix mileage field type from Int to Long to support larger values
The mileage field in TechnicalStateRequest was defined as Int, which has
a maximum value of 2,147,483,647. This caused deserialization errors when
mileage values exceeded this limit (e.g., 4,234,243,242).

Changes:
- Changed TechnicalStateRequest.mileage from Int to Long
- Removed unnecessary .toLong() conversion in CreateVisitFromReservationHandler
  since mileage is now already a Long

This fix allows the API to accept mileage values up to 9,223,372,036,854,775,807
(Long.MAX_VALUE), which is sufficient for any realistic vehicle mileage.

+++ src/main/kotlin/pl/detailing/crm/checkin/CheckinController.kt
@@ -136,7 +136,7 @@ data class VehicleRequest(
 )
 
 data class TechnicalStateRequest(
-    val mileage: Int,
+    val mileage: Long,
     val deposit: DepositItemRequest,
     val inspectionNotes: String
 )
